Almost finished

+++ pages/api/seed/index.ts
@@ -1,23 +1,23 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { db } from '../../../database';
-import { allProducts } from '../../../interfaces';
-import { Product } from '../../../models';
+import { allProducts, adoptionPets, otherPets } from '../../../interfaces';
+import { Pet, Product } from '../../../models';
 
 type Data =
 | { error: boolean; message: string }
 
 const validSeeds = [
     'products',
-    'animales',
+    'pets',
 ]
 
 export default function handler (req: NextApiRequest, res: NextApiResponse<Data>) {
     
     if ( req.method !== 'GET' ) return res.status(400).json({ error: true, message: "BAD REQUEST" });
-    if ( !req.query.seed ) return res.status(400).json({ error: true, message: 'Debes indicar lo que vas a cargar a la DB' });
-    if ( !validSeeds.includes( req.query.seed.toString() ) ) return res.status(400).json({ error: true, message: 'Esa query no existe' });
+    if ( !req.query.data ) return res.status(400).json({ error: true, message: 'Debes indicar lo que vas a cargar a la DB' });
+    if ( !validSeeds.includes( req.query.data.toString() ) ) return res.status(400).json({ error: true, message: 'Esa query no existe' });
 
-    if ( req.query.seed === 'products' ) return updateProducts( req, res );
+    if ( req.query.data ) return updateProducts( req, res );
 
     return res.status(201).json({ error: false, message: 'No hay data' })
 
@@ -25,25 +25,51 @@ export default function handler (req: NextApiRequest, res: NextApiResponse<Data>
 
 const updateProducts = async ( req: NextApiRequest, res: NextApiResponse ) => {
 
+    const { data } = req.query;
+
     try {
         await db.connect();
-        
-        await Product.deleteMany();
 
-        const actualProducts = allProducts.map(( product ) => {
-            let { _id, ...p } = product;
-            return p;
-        })
+        if ( data === 'products' ) {
+            await Product.deleteMany();
+    
+            const actualProducts = allProducts.map(( product ) => {
+                let { _id, ...p } = product;
+                return p;
+            });
+            
+            await Product.insertMany( actualProducts );
+
+            await db.disconnect();
+
+            return res.status(200).json({ error: false, message: 'Los productos fueron cargados' });
+        }
         
-        await Product.insertMany( actualProducts );
+        if ( data === 'pets' ) {
+            await Pet.deleteMany();
+    
+            const actualPets = adoptionPets.map(( pet ) => {
+                let { _id, ...p } = pet;
+                return p;
+            });
+
+            const morePets = otherPets.map(( pet ) => {
+                let { _id, ...p } = pet;
+                return p;
+            });
+            
+            await Pet.insertMany([...actualPets, ...morePets]);
+
+            await db.disconnect();
+
+            return res.status(200).json({ error: false, message: 'Las mascotas fueron cargadas' });
+        }
         
-        await db.disconnect();
-        
-        return res.status(200).json({ error: false, message: 'Los productos fueron cargados' });
+        return res.status(200).json({ error: false, message: 'Oops, la request pasó' });
     } catch( error ) {
         console.log( error );
         await db.disconnect();
-        return res.status(500).json({ error: true, message: "Updating products failed. Check server's logs" });
+        return res.status(500).json({ error: true, message: "Updating pets failed. Check server's logs" });
     }
 
 }
